% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{addFilesToDataset}
\alias{addFilesToDataset}
\title{Add Files to Dataset}
\usage{
addFilesToDataset(datasetID, fileData)
}
\arguments{
\item{datasetID}{The dataset ID.}

\item{fileData}{A named list of file metadata.}
}
\value{
A JSON object confirming file addition.
}
\description{
Adds file metadata to a dataset.
}
