% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{authenticate}
\alias{authenticate}
\title{Authenticate with the `Via Foundry` API}
\usage{
authenticate(
  hostname,
  username = NULL,
  password,
  identity_type = 1,
  redirect_uri = "http://localhost"
)
}
\arguments{
\item{hostname}{The API `hostname` (e.g., `https://your_viafoundry`).}

\item{username}{The login username.}

\item{password}{The login password.}

\item{identity_type}{The identity type (default is 1).}

\item{redirect_uri}{The redirect `URI` (default is `http://your_viafoundry/redirect`).}
}
\value{
None. Saves the authentication token to the configuration file.
}
\description{
Authenticates the user with the `Via Foundry` API using their username and password.
Stores the hostname and authentication token in a local configuration file.
}
\examples{
\dontrun{
authenticate("https://your_viafoundry", "username", "password")
}
}
