% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{calculate_expiration_date}
\alias{calculate_expiration_date}
\title{Calculate expiration date (30 days from now)}
\usage{
calculate_expiration_date()
}
\value{
The expiration date as a string.
}
\description{
Calculate expiration date (30 days from now)
}
