% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proccess.R
\name{checkProcessUsage}
\alias{checkProcessUsage}
\title{Check Process Usage}
\usage{
checkProcessUsage(processID)
}
\arguments{
\item{processID}{The ID of the process.}
}
\value{
A JSON object indicating usage information.
}
\description{
Checks if a process is used in pipelines or runs.
}
