% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{createCanvas}
\alias{createCanvas}
\title{Create Canvas}
\usage{
createCanvas(canvasData)
}
\arguments{
\item{canvasData}{A named list with canvas details.}
}
\value{
A JSON object with the created canvas.
}
\description{
Create Canvas
}
\examples{
\dontrun{
canvasData <- list(
  name = "new_canvas",
  label = "New canvas"
)
createCanvas(canvasData)
}
}
