% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{createCollection}
\alias{createCollection}
\title{Create Collection}
\usage{
createCollection(collectionData)
}
\arguments{
\item{collectionData}{A named list with collection details.}
}
\value{
A JSON object with the created collection.
}
\description{
Create Collection
}
