% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{createField}
\alias{createField}
\title{Create Field}
\usage{
createField(fieldData)
}
\arguments{
\item{fieldData}{A named list with field details.}
}
\value{
A JSON object with the created field.
}
\description{
Create Field
}
