% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proccess.R
\name{createMenuGroup}
\alias{createMenuGroup}
\title{Create Menu Group}
\usage{
createMenuGroup(name)
}
\arguments{
\item{name}{The name of the menu group to create.}
}
\value{
A JSON object containing the created menu group details.
}
\description{
Creates a new menu group.
}
