% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proccess.R
\name{createParameter}
\alias{createParameter}
\title{Create a Parameter}
\usage{
createParameter(parameterData)
}
\arguments{
\item{parameterData}{A list containing the parameter data.}
}
\value{
A JSON object containing the created parameter details.
}
\description{
Creates a new parameter with the given data.
}
