% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proccess.R
\name{createProcess}
\alias{createProcess}
\title{Create a New Process}
\usage{
createProcess(processData)
}
\arguments{
\item{processData}{A list containing the process data.}
}
\value{
A JSON object containing the created process details.
}
\description{
Creates a new process using the provided data.
}
