% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proccess.R
\name{createProcessConfig}
\alias{createProcessConfig}
\title{Create Process Config}
\usage{
createProcessConfig(
  name,
  menuGroupName,
  inputParams,
  outputParams,
  summary = "",
  scriptBody = "",
  scriptLanguage = "bash",
  scriptHeader = "",
  scriptFooter = "",
  permissionSettings = list(viewPermissions = 3, writeGroupIds = list()),
  revisionComment = "Initial revision"
)
}
\arguments{
\item{name}{Name of the process.}

\item{menuGroupName}{Name of the menu group.}

\item{inputParams}{A list of input parameter specs.}

\item{outputParams}{A list of output parameter specs.}

\item{summary}{Optional. Summary string.}

\item{scriptBody}{Optional. Script body.}

\item{scriptLanguage}{Optional. Language (default "bash").}

\item{scriptHeader}{Optional. Script header.}

\item{scriptFooter}{Optional. Script footer.}

\item{permissionSettings}{Optional. A list of permission settings.}

\item{revisionComment}{Optional. Revision comment.}
}
\value{
A list representing the full process config.
}
\description{
Assembles a complete process configuration object.
}
