% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{deleteCanvas}
\alias{deleteCanvas}
\title{Delete canvas}
\usage{
deleteCanvas(canvasID)
}
\arguments{
\item{canvasID}{A character string with canvas ID.}
}
\value{
A confirmation object or message.
}
\description{
Delete canvas
}
\examples{
 \dontrun{
canvasData <- list(
  name = "new_canvas",
  label = "New canvas"
)
createCanvas(canvasData)
search_params <- list(
filter = list(
 name="new_canvas"
)
)
canvas <- searchCanvas(search_params)
deleteCanvas(canvas$data$`_id`[1])
}
}
