% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{deleteCollection}
\alias{deleteCollection}
\title{Delete Collection}
\usage{
deleteCollection(collectionID)
}
\arguments{
\item{collectionID}{A character string with collection ID.}
}
\value{
A confirmation object or message.
}
\description{
Delete Collection
}
