% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{deleteData}
\alias{deleteData}
\title{Delete Data Entry}
\usage{
deleteData(canvasID, collectionName, dataID)
}
\arguments{
\item{canvasID}{The canvas ID.}

\item{collectionName}{The collection name.}

\item{dataID}{The ID of the data entry.}
}
\value{
A JSON confirmation.
}
\description{
Deletes a data entry by ID.
}
