% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{deleteField}
\alias{deleteField}
\title{Delete Field}
\usage{
deleteField(fieldID)
}
\arguments{
\item{fieldID}{A character string with field ID.}
}
\value{
A confirmation object or message.
}
\description{
Delete Field
}
