% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proccess.R
\name{deleteParameter}
\alias{deleteParameter}
\title{Delete a Parameter}
\usage{
deleteParameter(parameterID)
}
\arguments{
\item{parameterID}{The ID of the parameter to delete.}
}
\value{
A confirmation message.
}
\description{
Deletes a specific parameter by its ID.
}
