% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proccess.R
\name{deleteProcess}
\alias{deleteProcess}
\title{Delete a Process}
\usage{
deleteProcess(processID)
}
\arguments{
\item{processID}{The ID of the process to delete.}
}
\value{
A confirmation message.
}
\description{
Deletes a specific process by its ID.
}
