% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{discover}
\alias{discover}
\title{Discover Available Endpoints}
\usage{
discover()
}
\value{
A character vector of available endpoints.
}
\description{
Fetches and lists all available API endpoints from the Swagger documentation.
}
\examples{
\dontrun{
discover()
}
}
