% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proccess.R
\name{duplicateProcess}
\alias{duplicateProcess}
\title{Duplicate Process}
\usage{
duplicateProcess(processID)
}
\arguments{
\item{processID}{The ID of the process to duplicate.}
}
\value{
A JSON object containing the duplicated process details.
}
\description{
Creates a duplicate of an existing process.
}
