% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{fetchReportData}
\alias{fetchReportData}
\title{Fetch the JSON data for a report}
\usage{
fetchReportData(reportID)
}
\arguments{
\item{reportID}{The ID of the report to fetch data for.}
}
\value{
The JSON object containing the report data.
}
\description{
Fetch the JSON data for a report
}
