% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proccess.R
\name{filterParameters}
\alias{filterParameters}
\title{Filter Parameters}
\usage{
filterParameters(name = NULL, qualifier = NULL, fileType = NULL, id = NULL)
}
\arguments{
\item{name}{Optional. Name to filter.}

\item{qualifier}{Optional. Qualifier to filter.}

\item{fileType}{Optional. File type to filter.}

\item{id}{Optional. ID to filter.}
}
\value{
A list of filtered parameters.
}
\description{
Filters parameters by name, qualifier, file type, and/or ID.
}
