% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{get_api_status}
\alias{get_api_status}
\title{Get API Status}
\usage{
get_api_status()
}
\value{
A character vector with the API status.
}
\description{
Sends a simple GET request to check the status of the API.
}
\examples{
\dontrun{
status <- get_api_status()
print(status)
}
}
