% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_bearer_token}
\alias{get_bearer_token}
\title{Get bearer token using the cookie token}
\usage{
get_bearer_token(hostname, cookie_token, name = "token")
}
\arguments{
\item{hostname}{The API url}

\item{cookie_token}{The cookie token.}

\item{name}{The name of the token (default is "token").}
}
\value{
The bearer token.
}
\description{
Get bearer token using the cookie token
}
