% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_headers}
\alias{get_headers}
\title{Get headers for API requests}
\usage{
get_headers(config_path = DEFAULT_CONFIG_PATH)
}
\arguments{
\item{config_path}{The path to the configuration file (default is `~/.Rviaenv`).}
}
\value{
A list of headers with the bearer token.
}
\description{
Get headers for API requests
}
