% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{getAllReportPaths}
\alias{getAllReportPaths}
\title{Get unique report directories and attempt IDs for a specific report}
\usage{
getAllReportPaths(report_id)
}
\arguments{
\item{report_id}{The ID of the report.}
}
\value{
A character vector of unique report directories.
}
\description{
Get unique report directories and attempt IDs for a specific report
}
