% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{getCanvas}
\alias{getCanvas}
\title{Get canvas}
\usage{
getCanvas(canvasID)
}
\arguments{
\item{canvasID}{A character string with canvas ID.}
}
\value{
A JSON object with canvas details.
}
\description{
Get canvas
}
\examples{
\dontrun{
canvasData <- list(
  name = "new_canvas",
  label = "New canvas"
)
createCanvas(canvasData)
search_params <- list(
filter = list(
 name="new_canvas"
)
)
canvas <- searchCanvas(search_params)
getCanvas(canvas$data$`_id`[1])

}
}
