% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{getCanvasFields}
\alias{getCanvasFields}
\title{Get canvas Fields}
\usage{
getCanvasFields(canvasID)
}
\arguments{
\item{canvasID}{The ID of the canvas.}
}
\value{
A JSON object with a list of all fields.
}
\description{
Retrieves all metadata fields for a given canvas by collecting them from each collection.
}
