% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{getCollection}
\alias{getCollection}
\title{Get Collection}
\usage{
getCollection(collectionID)
}
\arguments{
\item{collectionID}{A character string with collection ID.}
}
\value{
A JSON object with collection details.
}
\description{
Get Collection
}
