% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{getCollectionFields}
\alias{getCollectionFields}
\title{Get Collection Fields}
\usage{
getCollectionFields(collectionID)
}
\arguments{
\item{collectionID}{The ID of the collection.}
}
\value{
A JSON object with a list of fields.
}
\description{
Retrieves metadata fields by collection ID.
}
