% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{getData}
\alias{getData}
\title{Get Data Entry}
\usage{
getData(canvasID, collectionName, dataID)
}
\arguments{
\item{canvasID}{The canvas ID.}

\item{collectionName}{The collection name.}

\item{dataID}{The ID of the data record.}
}
\value{
A JSON object with data entry.
}
\description{
Retrieves a single data record by ID.
}
