% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{getField}
\alias{getField}
\title{Get Field}
\usage{
getField(fieldID)
}
\arguments{
\item{fieldID}{A character string with field ID.}
}
\value{
A JSON object with field details.
}
\description{
Get Field
}
