% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{getFieldsForCollection}
\alias{getFieldsForCollection}
\title{Get Fields for a Specific Collection in a canvas}
\usage{
getFieldsForCollection(canvasID, collectionName)
}
\arguments{
\item{canvasID}{The ID of the canvas.}

\item{collectionName}{The name of the collection (e.g., "surveys").}
}
\value{
A list of field definitions for the collection.
}
\description{
Given a canvas ID and collection name, returns the metadata fields defined for that collection.
}
