% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proccess.R
\name{getMenuGroupByName}
\alias{getMenuGroupByName}
\title{Get Menu Group by Name}
\usage{
getMenuGroupByName(groupName)
}
\arguments{
\item{groupName}{The name of the menu group.}
}
\value{
The ID of the menu group if found, otherwise NULL.
}
\description{
Searches for a menu group by name and returns its ID.
}
