% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proccess.R
\name{getPipelineParameters}
\alias{getPipelineParameters}
\title{Get Pipeline Parameters}
\usage{
getPipelineParameters(pipelineID)
}
\arguments{
\item{pipelineID}{The ID of the pipeline.}
}
\value{
A JSON object containing the list of parameters.
}
\description{
Retrieves the list of parameters for a specific pipeline.
}
