% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proccess.R
\name{getProcess}
\alias{getProcess}
\title{Get Process Information}
\usage{
getProcess(processID)
}
\arguments{
\item{processID}{The ID of the process to retrieve.}
}
\value{
A JSON object containing the process details.
}
\description{
Fetches detailed information about a specific process.
}
