% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{getProcessNames}
\alias{getProcessNames}
\title{Get unique process names}
\usage{
getProcessNames(json_data)
}
\arguments{
\item{json_data}{The JSON object containing the report data.}
}
\value{
A character vector of unique process names.
}
\description{
Get unique process names
}
