% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proccess.R
\name{getProcessRevisions}
\alias{getProcessRevisions}
\title{Get Process Revisions}
\usage{
getProcessRevisions(processID)
}
\arguments{
\item{processID}{The ID of the process.}
}
\value{
A JSON object containing the revisions.
}
\description{
Fetches all revisions for a given process.
}
