% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{getReportDirs}
\alias{getReportDirs}
\title{Get directories following pub web in the route path}
\usage{
getReportDirs(report_id)
}
\arguments{
\item{report_id}{The ID of the report.}
}
\value{
A character vector of unique directories found after pub web.
}
\description{
Get directories following pub web in the route path
}
