% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proccess.R
\name{listMenuGroups}
\alias{listMenuGroups}
\title{List Menu Groups}
\usage{
listMenuGroups()
}
\value{
A JSON object containing the list of menu groups.
}
\description{
Fetches all menu groups from the API.
}
