% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proccess.R
\name{listParameters}
\alias{listParameters}
\title{List All Parameters}
\usage{
listParameters()
}
\value{
A JSON object containing the list of parameters.
}
\description{
Fetches all parameters from the API.
}
