% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proccess.R
\name{listProcesses}
\alias{listProcesses}
\title{List All Processes}
\usage{
listProcesses()
}
\value{
A JSON object containing the list of processes.
}
\description{
Fetches all existing processes from the API.
}
