% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{load_config}
\alias{load_config}
\title{Load Configuration}
\usage{
load_config()
}
\value{
A list containing the host name and bearer token.
}
\description{
Loads the configuration file or triggers authentication if the file is missing.
}
