% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{loadFile}
\alias{loadFile}
\title{Load or download a file from a process and file name}
\usage{
loadFile(json_data, processName, fileName, sep = "\\t", download_dir = getwd())
}
\arguments{
\item{json_data}{The `JSON` object containing the report data.}

\item{processName}{The name of the process.}

\item{fileName}{The name of the file to load or download.}

\item{sep}{The separator for tabular files. Default is tab-separated.}

\item{download_dir}{The directory where non-tabular files will be downloaded.}
}
\value{
A data frame with the file contents if the file is tabular; otherwise, NULL after downloading the file.
}
\description{
Load or download a file from a process and file name
}
