% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{searchCanvas}
\alias{searchCanvas}
\title{Search canvas}
\usage{
searchCanvas(searchParams = list(filter = list()))
}
\arguments{
\item{searchParams}{A named list of search filters.}
}
\value{
A JSON object with canvas search results.
}
\description{
Search canvas
}
\examples{
 \dontrun{
search_params <- list(
  filter = list(
    name="new_canvas"
  )
)
canvas <- searchCanvas(search_params)
print(canvas)
}
}
