% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{searchCollections}
\alias{searchCollections}
\title{Search Collections}
\usage{
searchCollections(searchParams = list(filter = list()))
}
\arguments{
\item{searchParams}{A named list of search filters.}
}
\value{
A JSON object with collection search results.
}
\description{
Search Collections
}
