% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{searchData}
\alias{searchData}
\title{Search Collection Data}
\usage{
searchData(canvasID, collectionName, filterData = list(filter = list()))
}
\arguments{
\item{canvasID}{The canvas ID.}

\item{collectionName}{The name of the collection.}

\item{filterData}{Optional filters.}
}
\value{
A JSON object of matching data entries.
}
\description{
Searches entries in a metadata collection.
}
