% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{searchDatasetFiles}
\alias{searchDatasetFiles}
\title{Search Dataset Files}
\usage{
searchDatasetFiles(datasetID, filterData = list(filter = list()))
}
\arguments{
\item{datasetID}{The ID of the dataset.}

\item{filterData}{A named list of filter parameters.}
}
\value{
A JSON object of matching files.
}
\description{
Lists files associated with a dataset.
}
