% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{searchFields}
\alias{searchFields}
\title{Search Fields}
\usage{
searchFields(searchParams = list(filter = list()))
}
\arguments{
\item{searchParams}{A named list of search filters.}
}
\value{
A JSON object with field search results.
}
\description{
Search Fields
}
