% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{updateCanvas}
\alias{updateCanvas}
\title{Update canvas}
\usage{
updateCanvas(canvasID, updateData)
}
\arguments{
\item{canvasID}{A character string with canvas ID.}

\item{updateData}{A named list of canvas updates.}
}
\value{
A JSON object with updated canvas.
}
\description{
Update canvas
}
