% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{updateCollection}
\alias{updateCollection}
\title{Update Collection}
\usage{
updateCollection(collectionID, updateData)
}
\arguments{
\item{collectionID}{A character string with collection ID.}

\item{updateData}{A named list of collection updates.}
}
\value{
A JSON object with updated collection.
}
\description{
Update Collection
}
