% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{updateData}
\alias{updateData}
\title{Update Data Entry}
\usage{
updateData(canvasID, collectionName, dataID, updateData)
}
\arguments{
\item{canvasID}{The canvas ID.}

\item{collectionName}{The collection name.}

\item{dataID}{The ID of the data to update.}

\item{updateData}{A named list of updates.}
}
\value{
A JSON object with updated entry.
}
\description{
Updates an existing data record.
}
