% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{updateField}
\alias{updateField}
\title{Update Field}
\usage{
updateField(fieldID, updateData)
}
\arguments{
\item{fieldID}{A character string with field ID.}

\item{updateData}{A named list of field updates.}
}
\value{
A JSON object with updated field.
}
\description{
Update Field
}
