% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proccess.R
\name{updateMenuGroup}
\alias{updateMenuGroup}
\title{Update Menu Group}
\usage{
updateMenuGroup(menuGroupID, name)
}
\arguments{
\item{menuGroupID}{The ID of the menu group to update.}

\item{name}{The new name of the menu group.}
}
\value{
A JSON object containing the updated menu group details.
}
\description{
Updates the name of an existing menu group.
}
