% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proccess.R
\name{updateParameter}
\alias{updateParameter}
\title{Update a Parameter}
\usage{
updateParameter(parameterID, parameterData)
}
\arguments{
\item{parameterID}{The ID of the parameter to update.}

\item{parameterData}{A list containing the updated parameter data.}
}
\value{
A JSON object containing the updated parameter details.
}
\description{
Updates an existing parameter with the provided data.
}
