% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proccess.R
\name{updateProcess}
\alias{updateProcess}
\title{Update a Process}
\usage{
updateProcess(processID, processData)
}
\arguments{
\item{processID}{The ID of the process to update.}

\item{processData}{A list containing the updated process data.}
}
\value{
A JSON object containing the updated process details.
}
\description{
Updates an existing process with the given data.
}
