% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{uploadReportFile}
\alias{uploadReportFile}
\title{Upload a file to a specific report}
\usage{
uploadReportFile(report_id, file_path, dir = NULL)
}
\arguments{
\item{report_id}{The ID of the report for the API.}

\item{file_path}{The local path to the file being uploaded.}

\item{dir}{An optional directory name for organizing files.}
}
\value{
A list containing the server response.
}
\description{
Upload a file to a specific report
}
