% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{getAllFileNames}
\alias{getAllFileNames}
\title{Extract children names across all processes}
\usage{
getAllFileNames(json_data)
}
\arguments{
\item{json_data}{The `JSON` object containing the report data.}
}
\value{
A data frame containing `id`, `processName`, `name`, `extension`, `fileSize`, and `routePath`.
}
\description{
Extract children names across all processes
}
