% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{getAllFileNames}
\alias{getAllFileNames}
\title{Recursively extract all files from deeply nested report JSON}
\usage{
getAllFileNames(json_data)
}
\arguments{
\item{json_data}{The parsed report JSON returned from fetchReportData()}
}
\value{
A data frame of all files and their metadata
}
\description{
Recursively extract all files from deeply nested report JSON
}
