% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{getFileNames}
\alias{getFileNames}
\title{Get file names for a specific process}
\usage{
getFileNames(json_data, processName)
}
\arguments{
\item{json_data}{The `JSON` object containing the report data.}

\item{processName}{The name of the process to filter by.}
}
\value{
A data frame containing `id`, `name`, `extension`, `fileSize`, and `routePath`.
}
\description{
Get file names for a specific process
}
