% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{getReportDirs}
\alias{getReportDirs}
\title{Get directories following "pubweb" in the routePath}
\usage{
getReportDirs(report_id)
}
\arguments{
\item{report_id}{The ID of the report.}
}
\value{
A character vector of unique directories found after "pubweb".
}
\description{
Get directories following "pubweb" in the routePath
}
