% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{login}
\alias{login}
\title{Login and retrieve the cookie token}
\usage{
login(
  hostname,
  username,
  password,
  identity_type = 1,
  redirect_uri = "http://localhost"
)
}
\arguments{
\item{hostname}{The API hostname.}

\item{username}{The login username.}

\item{password}{The login password.}

\item{identity_type}{The identity type.}

\item{redirect_uri}{The redirect URI.}
}
\value{
The cookie token.
}
\description{
Login and retrieve the cookie token
}
